fix: change code directory architecture to fit product convention
