--- models.py
@@ -1,10 +1,0 @@
-from pydantic import BaseModel
-from typing import List 
-
-
-class Recipe(BaseModel):
-    name: str
-    type: str
-    difficulty: int
-    ingredients: List[str]
-    preparation: str
