+++ recipe/routes/recipe.py
@@ -1,4 +1,6 @@
-from fastapi import APIRouter, HTTPException
+from typing import Optional
+from fastapi import APIRouter, HTTPException, Query
+from models import Recipe
 from recipe.models.recipe import Item
 from core.database import db
 from bson import ObjectId
@@ -8,15 +10,31 @@ router = APIRouter(
     tags=["Recipe"],
 )
 
-@router.post("/items/")
-async def create_item(item: Item):
-    result = await db["items"].insert_one(item.dict())
-    return {"id": str(result.inserted_id)}
+@router.post("/createRecipes/")
+async def create_recipe(recipe: Recipe):
+    if db is None:
+        raise HTTPException(status_code=500, detail="Database connection error")
+    recipes_collection = db["recipes"]
 
-@router.get("/items/")
-async def list_items():
-    items = await db["items"].find().to_list(100)
-    return [
-        {"id": str(item["_id"]), **{k: v for k, v in item.items() if k != "_id"}}
-        for item in items
-    ]
+    if recipes_collection.find_one({"name": recipe.name}):
+        raise HTTPException(status_code=400, detail="Recipe with this name already exists")
+    
+    recipes_collection.insert_one(recipe.model_dump())
+    return {"message": "Recipe created successfully", "recipe": recipe}
+
+
+@router.get("/getRecipes")
+async def get_recipes(name: Optional[str] = Query(None, description="Optional name filter for recipes")):
+    if db is None:
+        raise HTTPException(status_code=500, detail="Database connection error")
+
+    recipes_collection = db["recipes"]
+
+    query = {"name": name} if name else {}
+
+    recipes = list(recipes_collection.find(query))
+
+    for recipe in recipes:
+        recipe["_id"] = str(recipe["_id"])
+
+    return recipes
